Replace kotlin-extensions and kotlinx-nodejs by kotlin-wrappers

+++ gradle/build-logic/src/main/kotlin/net/kautler/node.gradle.kts
@@ -130,8 +130,9 @@ configure<NodeJsRootExtension> {
 
 dependencies {
     implementation(libs.kotlinx.coroutines.core)
-    implementation(libs.kotlin.extensions)
-    implementation(libs.kotlinx.nodejs)
+    implementation(platform(libs.kotlin.wrappers.bom))
+    implementation(libs.kotlin.wrapper.js)
+    implementation(libs.kotlin.wrapper.node)
     implementation(npm(libs.actions.cache))
     implementation(npm(libs.actions.core))
     implementation(npm(libs.actions.exec))
@@ -163,22 +164,55 @@ tasks.withType(IntegratedDukatTask::class).configureEach {
         )
         fixExternalsFiles(
             this,
+            "index.module_@actions_http-client.kt" to listOf(
+                """\Qimport http.IncomingMessage\E$""" to "import node.http.IncomingMessage",
+                """\Qimport http.OutgoingHttpHeaders\E$""" to "import node.http.OutgoingHttpHeaders",
+                """\Qimport NodeJS.ReadableStream\E$""" to "import node.ReadableStream",
+                """\Qimport http.Agent\E$""" to "import node.http.Agent",
+                """\Qtypealias HttpClientError = Error\E$""" to "external class HttpClientError : Throwable"
+            ),
+            "interfaces.module_@actions_exec.kt" to listOf(
+                """\Qimport stream.internal.Writable\E$""" to "import node.stream.Writable",
+                """\Qimport buffer.global.Buffer\E$""" to "import node.buffer.Buffer",
+                """\Qvar env: `T$2`?\E\r?\n\Q        get() = definedExternally\E\r?\n\Q        set(value) = definedExternally\E""" to "var env: js.core.ReadonlyRecord<String, String>?"
+            ),
+            "interfaces.module_@actions_http-client.kt" to listOf(
+                """\Qimport http.OutgoingHttpHeaders\E$""" to "import node.http.OutgoingHttpHeaders",
+                """\Qimport NodeJS.ReadableStream\E$""" to "import node.ReadableStream",
+                """\Qimport http.IncomingHttpHeaders\E$""" to "import node.http.IncomingHttpHeaders",
+                """\Qexternal interface HttpClient {\E$""" to "external interface HttpClient2 {",
+                """\Qoptions: http.RequestOptions\E(\))?$""" to "options: node.http.RequestOptions$1"
+            ),
             "lib.dom.kt" to listOf(
-                """\Qoverride fun addEventListener(type: String, listener: EventListenerObject\E""" to """fun addEventListener(type: String, listener: EventListenerObject""",
-                """\Qoverride fun removeEventListener(type: String, callback: EventListenerObject\E""" to """fun removeEventListener(type: String, callback: EventListenerObject""",
-                """\Q`T${'$'}16`\E""" to """`T\${'$'}16`<R, T>"""
+                """\Qimport url.URL as _URL\E$""" to "import node.url.URL as _URL",
+                """\Qimport url.URLSearchParams as _URLSearchParams\E$""" to "import node.url.URLSearchParams as _URLSearchParams",
+                """\Qoverride fun addEventListener(type: String, listener: EventListenerObject\E""" to "fun addEventListener(type: String, listener: EventListenerObject",
+                """\Qoverride fun removeEventListener(type: String, callback: EventListenerObject\E""" to "fun removeEventListener(type: String, callback: EventListenerObject",
+                """\Q`T$16`\E""" to """`T\$16`<R, T>"""
             ),
             // work-around for https://github.com/Kotlin/dukat/issues/402
             "lib.es2018.asynciterable.module_dukat.kt" to listOf(
                 """\Qval `return`: ((value: TReturn) -> Promise<dynamic /* IteratorYieldResult<T> | IteratorReturnResult<TReturn> */>)?\E$""" to "val `return`: ((value: dynamic) -> Promise<dynamic /* IteratorYieldResult<T> | IteratorReturnResult<TReturn> */>)?",
-                """^*\Qval `return`: ((value: PromiseLike<TReturn>) -> Promise<dynamic /* IteratorYieldResult<T> | IteratorReturnResult<TReturn> */>)?\E\r?\n\Q        get() = definedExternally\E\r?\n""" to ""
+                """\Qval `return`: ((value: PromiseLike<TReturn>) -> Promise<dynamic /* IteratorYieldResult<T> | IteratorReturnResult<TReturn> */>)?\E\r?\n\Q        get() = definedExternally\E\r?\n""" to ""
+            ),
+            "lib.es2020.bigint.module_dukat.kt" to listOf(
+                """\Q : RelativeIndexable<Any>\E""" to ""
+            ),
+            "lib.es5.kt" to listOf(
+                """\Qimport NodeJS.CallSite\E$""" to "import node.CallSite",
+                """\Qval resolve: ((specified: String, parent: URL) -> Promise<String>)?\E$""" to "val resolve2: ((specified: String, parent: URL) -> Promise<String>)?"
             ),
             // work-around for https://github.com/Kotlin/dukat/issues/401
             "null-writable.module_null-writable.kt" to listOf(
-                """\Q`T$13`\E""" to """`T\$10`"""
+                """\Qimport stream.internal.`T$13`\E$""" to "",
+                """\Qimport stream.internal.Writable\E$""" to "import node.stream.Writable",
+                """\Qoverride fun _write(_chunk: Any, _encoding: String, callback: (error: Error?) -> Unit)\E$""" to "override fun _write(chunk: Any, encoding: node.buffer.BufferEncoding, callback: (error: Error?) -> Unit)",
+                """\Qopen fun _writev(_chunks: Array<`T$88`>, callback: (error: Error?) -> Unit)\E$""" to "",
+                """\Qoverride fun _writev(chunks: Array<`T$13`>, callback: (error: Error?) -> Unit)\E$""" to ""
             ),
             // work-around for https://github.com/Kotlin/dukat/issues/399
             "tool-cache.module_@actions_tool-cache.kt" to listOf(
+                """\Qimport http.OutgoingHttpHeaders\E$""" to "import node.http.OutgoingHttpHeaders",
                 """\Qtypealias HTTPError = Error\E$""" to "external class HTTPError : Throwable",
                 """\Qtypealias IToolRelease = IToolRelease\E$""" to "",
                 """\Qtypealias IToolReleaseFile = IToolReleaseFile\E$""" to ""
@@ -188,24 +222,9 @@ tasks.withType(IntegratedDukatTask::class).configureEach {
                 """\Qtypealias ValidationError = Error\E$""" to "external class ValidationError : Throwable",
                 """\Qtypealias ReserveCacheError = Error\E$""" to "external class ReserveCacheError : Throwable"
             ),
-            "index.module_@actions_http-client.kt" to listOf(
-                """\Qtypealias HttpClientError = Error\E$""" to "external class HttpClientError : Throwable"
-            ),
-            "interfaces.module_@actions_http-client.kt" to listOf(
-                """\Qexternal interface HttpClient {\E$""" to "external interface HttpClient2 {"
-            ),
             // work-around for https://github.com/Kotlin/dukat/issues/400
             "semver.module_semver.kt" to listOf(
                 """\Q@JsModule("semver")\E$""" to """@JsModule("semver/classes/semver")"""
-            ),
-            "interfaces.module_@actions_exec.kt" to listOf(
-                """\Qimport buffer.global.Buffer\E$""" to """import Buffer"""
-            ),
-            "lib.es5.kt" to listOf(
-                """\Qval resolve: ((specified: String, parent: URL) -> Promise<String>)?\E$""" to """val resolve2: ((specified: String, parent: URL) -> Promise<String>)?"""
-            ),
-            "lib.es2020.bigint.module_dukat.kt" to listOf(
-                """\Q : RelativeIndexable<Any>\E""" to """"""
             )
         )
     }
@@ -235,16 +254,18 @@ fun addJsModuleAnnotations(task: Task, vararg pairs: Pair<String, String>) {
 }
 
 fun deleteExternalsFiles(task: Task, vararg files: String) {
-    task
-        .outputs
-        .files
-        .asFileTree
-        .matching {
-            for (file in files) {
-                include("**/$file")
+    if (files.isNotEmpty()) {
+        task
+            .outputs
+            .files
+            .asFileTree
+            .matching {
+                for (file in files) {
+                    include("**/$file")
+                }
             }
-        }
-        .forEach { it.delete() }
+            .forEach { it.delete() }
+    }
 }
 
 fun fixExternalsFiles(task: Task, vararg pairs: Pair<String, List<Pair<String, String>>>) {
